Add a wrapper for execve(2)
luaposix exec and execp don't seem to offer a way to specify the
environment, so let's add a C module which wraps the system call and
implement niceties such as path searching in lua.

I'm not sure how the lib.freebsd.sys namespace should look, but what I
have now is awkward.